Добавлен вывод Integer и Float по захардкоженным именам

+++ LlvmHelper/LlvmHelper.h
@@ -11,7 +11,7 @@ public:
 	static llvm::Type * CreateType(llvm::LLVMContext & context, std::string const & type);
 	static llvm::Constant * CreateConstant(llvm::LLVMContext & context, std::string const & type, std::string const & value);
 	static llvm::Constant * CreateIntegerConstant(llvm::LLVMContext & context, int value);
-	static llvm::Constant * CreateFloatConstant(llvm::LLVMContext & context, float value);
+	static llvm::Constant * CreateFloatConstant(llvm::LLVMContext & context, double value);
 	static llvm::Value * CreateAdd(llvm::IRBuilder<> * builder, std::string const & type, llvm::Value * lhs, llvm::Value * rhs, std::string const & name);
 	static llvm::Value * CreateSub(llvm::IRBuilder<> * builder, std::string const & type, llvm::Value * lhs, llvm::Value * rhs, std::string const & name);
 	static llvm::Value * CreateMul(llvm::IRBuilder<> * builder, std::string const & type, llvm::Value * lhs, llvm::Value * rhs, std::string const & name);
